move some testing infrastructure to a common place
This will improve reusability with the upcoming unix sockets, which also
want to create test directories
