+++ glommio/src/io/dma_file.rs
@@ -330,26 +330,8 @@ impl DmaFile {
 #[cfg(test)]
 pub(crate) mod test {
     use super::*;
+    use crate::test_utils::*;
     use crate::{ByteSliceMutExt, Local};
-    use std::path::PathBuf;
-
-    #[derive(Copy, Clone)]
-    pub(crate) enum TestDirectoryKind {
-        TempFs,
-        PollMedia,
-        NonPollMedia,
-    }
-
-    pub(crate) struct TestDirectory {
-        pub(crate) path: PathBuf,
-        pub(crate) kind: TestDirectoryKind,
-    }
-
-    impl Drop for TestDirectory {
-        fn drop(&mut self) {
-            let _ = std::fs::remove_dir_all(&self.path);
-        }
-    }
 
     #[cfg(test)]
     pub(crate) fn make_test_directories(test_name: &str) -> std::vec::Vec<TestDirectory> {
@@ -367,32 +349,11 @@ pub(crate) mod test {
                 );
             }
             Ok(path) => {
-                let mut dir = PathBuf::from(path);
-                std::assert!(dir.exists());
-
-                dir.push(test_name);
-                let _ = std::fs::remove_dir_all(&dir);
-                std::fs::create_dir_all(&dir).unwrap();
-                vec.push(TestDirectory {
-                    path: dir,
-                    kind: TestDirectoryKind::PollMedia,
-                })
+                vec.push(make_poll_test_directory(path, test_name));
             }
         };
 
-        let mut dir = std::env::temp_dir();
-        dir.push(test_name);
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        let buf = statfs(&dir).unwrap();
-        let fstype = buf.filesystem_type();
-        let kind = if fstype == TMPFS_MAGIC {
-            TestDirectoryKind::TempFs
-        } else {
-            TestDirectoryKind::NonPollMedia
-        };
-
-        vec.push(TestDirectory { path: dir, kind });
+        vec.push(make_tmp_test_directory(test_name));
         return vec;
     }
